+++ src/attach/browser.js
@@ -1,2 +1,3 @@
+// @ts-check
 'use strict'
 module.exports = require('@rill/http/adapter/document').attach