+++ src/attach/index.js
@@ -1,3 +1,4 @@
+// @ts-check
 'use strict'
 // Attaching to the document is a noop server side.
-module.exports = function () {}
+module.exports = function (arg) {}