+++ src/request.js
@@ -1,3 +1,7 @@
+// @ts-check
+/** Type Definitions */
+/** @module rill/Request */
+/** @typedef {{ referrer?: string, referer?: string?, date?: string, host?: string, cookie?: string, 'user-agent'?: string, 'accept-language'?: string, connection?: string, 'cache-control'?: string, accept?: string, [x]: string|string[] }} Headers */
 'use strict'
 
 var URL = require('mini-url')
@@ -5,17 +9,15 @@ var QS = require('mini-querystring')
 var cookie = require('cookie')
 var toField = require('header-field')
 var URL_PARTS = URL.parts
-
-// Expose module.
-module.exports =
-Request['default'] = Request
+module.exports = Request['default'] = Request
 
 /**
  * Wrapper around nodejs `IncommingMessage` that has pre parsed url
  * and other conveinences.
  *
- * @param {Context} ctx - The context for the request.
- * @param {http.IncommingMessage} req - The original node request.
+ * @param {rill.Context} ctx - The context for the request.
+ * @param {http.IncomingMessage} req - The original node request.
+ *
  * @constructor
  */
 function Request (ctx, req) {
@@ -27,10 +29,23 @@ function Request (ctx, req) {
   var parsed = process.browser ? req._options.parsed : URL.parse(req.url, origin)
   this.ctx = ctx
   this.original = req
+  /** @type {string} */
   this.method = req.method
+  /** @type {Headers} */
   this.headers = req.headers
+  /** @type {object} */
   this.cookies = this.headers['cookie'] ? cookie.parse(this.headers['cookie']) : {}
+  /** @type {string} */
+  this.path = req.url
+  /** @type {object} */
   this.params = {}
+  /** @type {boolean} */
+  this.secure = secure
+  /** @type {string} */
+  this.origin = origin
+  /* istanbul ignore next */
+  /** @type {string} */
+  this.ip = (conn.remoteAddress || req.socket.remoteAddress || (conn.socket && conn.socket.remoteAddress))
 
   // Attach url parts.
   for (var part, i = URL_PARTS.length; i--;) {
@@ -38,15 +53,14 @@ function Request (ctx, req) {
     this[part] = parsed[part]
   }
 
-  this.path = req.url
-  this.secure = secure
-  this.origin = origin
+  /** @type {string} */
   this.matchPath = this.pathname
+  /** @type {string} */
   this.matchHost = this.hostname
+  /** @type {string[]} */
   this.subdomains = String(this.hostname).split('.').reverse().slice(2)
+  /** @type {object} */
   this.query = QS.parse(this.search, true)
-  /* istanbul ignore next */
-  this.ip = (conn.remoteAddress || req.socket.remoteAddress || (conn.socket && conn.socket.remoteAddress))
 }
 
 /**
@@ -56,7 +70,7 @@ function Request (ctx, req) {
  * request.get('Host') // -> 'test.com'
  *
  * @param {string} name - The header field to get from the request.
- * @return {string|string[]}
+ * @return {string|string[]|void}
  */
 Request.prototype.get = function (name) {
   return this.headers[toField(name)]