+++ test/response.test.js
@@ -10,7 +10,7 @@ describe('Response', function () {
       var request = agent(Rill()
         .use(respond(200, function (ctx) {
           ctx.res.cookie('a', 1, { httpOnly: true })
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -27,7 +27,7 @@ describe('Response', function () {
           ctx.res.clearCookie('a')
           unsetCookie = ctx.res.get('Set-Cookie')[0]
           ctx.res.status = 200
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -43,7 +43,7 @@ describe('Response', function () {
       var request = agent(Rill()
         .use(respond(200, function (ctx) {
           ctx.res.set('X-Custom-Header', 1)
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -59,7 +59,7 @@ describe('Response', function () {
           ctx.res.set('X-Custom-Header', 1)
           ctx.res.append('X-Custom-Header', 2)
           ctx.res.append('X-Custom-Header', 3)
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -74,7 +74,7 @@ describe('Response', function () {
         .use(respond(200, function (ctx) {
           ctx.res.set('X-Custom-Header', 1)
           assert.equal(ctx.res.get('X-Custom-Header'), 1)
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -90,7 +90,7 @@ describe('Response', function () {
           ctx.res.set('X-Custom-Header', 1)
           ctx.res.remove('X-Custom-Header')
           assert.equal(ctx.res.get('X-Custom-Header'), null)
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -103,7 +103,7 @@ describe('Response', function () {
       var request = agent(Rill()
         .use(function (ctx) {
           ctx.res.redirect('http://localhost/test')
-        }).listen())
+        }).listen().unref())
 
       return request
         .get('/')
@@ -115,7 +115,7 @@ describe('Response', function () {
         .use(function (ctx) {
           ctx.res.status = 301
           ctx.res.redirect('http://localhost/test')
-        }).listen())
+        }).listen().unref())
 
       return request
         .get('/')
@@ -127,7 +127,7 @@ describe('Response', function () {
       var request = agent(Rill()
         .use(function (ctx) {
           ctx.res.redirect('back')
-        }).listen())
+        }).listen().unref())
 
       return request
         .get('/')
@@ -143,7 +143,7 @@ describe('Response', function () {
           assert.throws(function () {
             ctx.res.redirect()
           }, TypeError, 'Rill#ctx.res.redirect: Cannot redirect, url not specified and alternative not provided.')
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')
@@ -153,22 +153,23 @@ describe('Response', function () {
 
   describe('#refresh', function () {
     it('should refresh the browser', function () {
-      var request = agent(Rill()
-        .use(respond(200, function (ctx) {
-          ctx.res.refresh()
-        })).listen({ port: 3001 }))
+      var server = Rill()
+        .use(respond(200, function (ctx) { ctx.res.refresh() }))
+        .listen()
+        .unref()
+      var request = agent(server)
 
       return request
         .get('/')
         .expect(200)
-        .expect('refresh', '0; url=http://127.0.0.1:3001/')
+        .expect('refresh', '0; url=http://127.0.0.1:' + server.address().port + '/')
     })
 
     it('should refresh to a referrer with "back"', function () {
       var request = agent(Rill()
         .use(respond(200, function (ctx) {
           ctx.res.refresh(5, 'back')
-        })).listen())
+        })).listen().unref())
 
       return request
         .get('/')